# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Queue Core API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Register",
                "operationId": "register",
                "requestBody": {
                    "description": "Authenticate",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "c_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authenticate"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "Authenticate",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile data",
                "operationId": "profile",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/patient/my_data": {
            "get": {
                "tags": [
                    "Data Patient"
                ],
                "summary": "Edit my data patient",
                "operationId": "getmydata",
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Data Patient"
                ],
                "summary": "Edit my data patient",
                "operationId": "editmydata",
                "requestBody": {
                    "description": "form",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "mother_name": {
                                        "type": "string"
                                    },
                                    "identity_number": {
                                        "type": "string"
                                    },
                                    "dob": {
                                        "type": "date"
                                    },
                                    "gender": {
                                        "description": "1 = female, 0 = male",
                                        "type": "int"
                                    },
                                    "blood_type": {
                                        "description": "ex: o",
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}